Fixed not published article test
The previous one was invalid, both as `article.pub_date` doesn't exist as a model field anymore, and as the previous `404` response status code was because the article was set to `hidden` - not only because its publication date/time was in the future. To make sure of this, I also added an extra test case to confirm that the same article is visible when its publication time is in the past.