Renamed files with - or _ to match convention
The most popular convention for naming CSS and JavaScript files seems to be using hyphens between words. In our codebase, however, the .css files have underscores instead of hyphens, to match the name of the template file they belong to. The .js files also have underscores, to simply match the naming convention of all the other files.