Added custom template for EditUsageRulesView
Now the view is not dependent on a completely unrelated template, which facilitates changing model-specific parts of the templates without affecting the other view. This does however duplicate most of the code of `contentbox/edit.html`, and the two templates should be reformatted in the future to reduce this duplication.