Removed max_length and added missing blank=True
In general, it's not a very good idea to limit the number of characters a user can input, as it's surprisingly often too short compared to real data. In addition to this, some SQL database systems (like SQLite and PostgreSQL) actually implement fixed-length and unlimited-length text/char fields the same way, so there isn't any performance impact choosing `TextField` over `CharField`.

Regarding `blank=True`, this argument should (in most cases) be added to fields that are either nullable or have a default value (the default value of `TextField`/`Charfield` is in practice always the empty string, even though it isn't technically the default value if not explicitly specified).

Also removed `null=True` on string-based fields (see https://docs.djangoproject.com/en/3.1/ref/models/fields/#null on why).

Added `UnlimitedCharField` (which is a `TextField` in disguise) and `URLTextField`, for convenience.