Fixed missing contents of previous migration
Added an `AlterField` for `Quota.machine_type` that removes `null=True` and `blank=True`, and for `MachineType.name` that removes `max_length`, both of which were missing in the last migration (`0015_machinetype`) - added in 4dc6e64787ca632bcfb0c63fc4de735605ebea2b.

Also corrected the type of `MachineType`'s `ordering` option from a list to a tuple in the migration that added it. This was also mistakenly not included in the commit mentioned above.