Made delete modal's prompt more verbose
Removing the `href` attribute from the `<a>` buttons also fixes the bug / unintended function of being able to e.g. (accidentally) middle-click one of these delete buttons, which would open a new tab with the URL of the delete form, which would crash the server, as no template is registered in these delete views. This latter issue can be fixed by returning 404 when trying to make a GET request to the delete URLs.