Fixed missing month translations in date widgets
Updating Fomantic-UI in the last commit involved overwriting the translations that were added manually to Fomantic-UI's source files (in 2318b405059a5ad76792db0b182109c462629d18). This was instead replaced with overriding the `months` and `monthsShort` attributes of Fomantic-UI's calendar. The mentioned manually added translations that were not related to month names, were not replaced, as I could not find any usages of them.