Fixed slightly invalid member tests
The assertions for the `disallowed_clients` in `_test_url_permissions()` did not really test whether they were allowed to send those requests, as the `status_code`s were actually `404`, because the redirected login URL could not be found when the `DEFAULT_HOST` was `"internal"`. Also, the `follow=True` argument would have followed any redirects to a login page, which would have resulted in a `200` `status_code`.