Renamed `email` path to `contact`
This is a more fitting name, as `email` could lead to some confusion, e.g. if it's referring to the `mail` app or other email functions, instead of just a list of email addresses.

Also made the path of `contact` be prefixed by `/about/`, so that the full path is now `/about/contact/`.

Note: This rename will require manually renaming the previous `email` ContentBox. I could have added a migration for this, but it seemed a bit unnecessary.