Fixed not being able to create reservation rule
This was caused by the `disabled` attribute, which made Django use the field's initial data (which was `None`) instead of the value passed by the view.

Also passed `machine_type.pk` instead of the whole object to the form, as that more closely matches the posted data (if `machine_type` had been part of the template).