Added redirects for old URLs
The URLs added were thought to have a high potential of being referenced other places on the internet, like in articles, posts on the Facebook page, and old course contracts.

The change from "/email/" to "/about/contact/" in the last commit was not added, as the "/email/" path is generic enough to possibly be used for other purposes in the future.