Made NewsBase abstract
When using SQLite, this migration should apply painlessly. If using any other database management system, one might have to split it into multiple migrations to make it work.

Also, I was unable to set the new `can_view_private` permissions on objects having the old permission (of the same codename) in the migration, so this will have to be updated manually.

Lastly, the changes in `test_utils.py` fixes a crash if multiple permissions shared the same `codename` - made apparent by the migration adding the `can_view_private` permission for both `Article` and `Event`.