Made `verbose_name`s start with lowercase letter
Also did the same with `ModelAdmin`s' `short_description`, and various other places where a "capped first" version of the lowercase `verbose_name`s were being used. For the latter case, `capfirst` was added - in both Python and template code, to not change how the templates are rendered.

This change enables Django to selectively make the first letter of fields' `verbose_name` uppercase when rendering a template, depending on where the `verbose_name` is used; see the desccription of the convention here: https://docs.djangoproject.com/en/3.1/topics/db/models/#verbose-field-names.
Another advantage of this is that it doesn't litter the translation file with upper- and lowercased versions of the same word.