Removed `ui` CSS class from icons
Specifying the `ui` class is not part of the specification (see https://fomantic-ui.com/elements/icon.html), and should be removed to not cause unintended side-effects, like the sticky note icon (on the "Admin page" button on the article and event detail pages) sticking to the top of the page when scrolling.

Also reordered the classes of some icon tags, to make more sense grammatically.