Renamed template and static files for consistency
Renaming `reservation_create.html` to `reservation_edit.html` does not really matter - as the template is used for both editing and creating, but it's more common that editing is the "base action" and creating is a special case of editing (where the object doesn't exist yet).

Also removed the linked CSS file from `equipment_detail.html`, as none of its styles were used, and removed the `og_description` block, as the text didn't really make sense for a page on the details of a piece of equipment.