Renamed Member email -> contact_email
This helps distinguishing the purpose of the field from `User.email`.

Also removed setting the (contact) email to the user's email when creating a Member object, as the user should be able to decide their own preferred contact email address instead of the code taking that choice for them. So if we don't have the contact email of a user, we should ask them instead of assuming.