Refactored ReservationRule.start_days
Instead of `start_days` being an int where the 7 least significant bits represent the weekdays, it's now an array of ints, where each element is the index of the day it represents (Monday=1; Sunday=7).
This makes it a lot easier to understand what the value stored in the database means, and it reduces the code necessary to convert between different representations.