Fixed requests not marked as secure on server
Previously, on the deployment server, the requests were always marked as insecure, resulting in e.g. `request.scheme` being "http".

This was motivated by Facebook not showing the image of an article that was embedded in a post
(https://www.facebook.com/MAKENTNU/posts/701447270512461), which was (most likely) caused by the `og:image` meta tag
(in the `<head>` tag of `base.html`) specifying an image URL with "http" instead of "https" (because of `request.scheme`, as mentioned above).