Fixed old page version warning for current content
Also renamed the `related_name` of `Page.current_content` to a more fitting name.

Also, made the current page (content) version on the history page link to the content's absolute (history) URL, instead of the page's URL,
as it enables users to get a "permalink" to that specific page version.