Fixed disappearing course registration rows
After a page refresh of the course registration list, without changing the sorting or filtering: If the order of the users from the template context
differed from the order that they would appear in after being sorted with the `sort()` function, the rows that were "out of place" would disappear
when selecting them (they would actually just "jump" to their proper place - determined by the `sort()` function - which often happened to be on
another list page). This has now been fixed by sorting the list on page load.