Replaced HTTP status codes with enums
Some of the numerical values were left alone, e.g. when comparing the status code of a response to whether it's greater than `300`, as these cases
really test whether the status code is a member of the 3xx class (or higher), and it could be more confusing to read e.g.
`status_code > HTTPStatus.MULTIPLE_CHOICES` compared to just `status_code > 300`.

Also changed the order of some of the arguments to `TestCase.assertEqual()`, so that the logged messages make more sense (when those tests fail).