Simplified machine's machine type widget
Also readded `machine_type` to the list of excluded fields in `EditMachineForm`, as the type of a machine should not be changed after creation,
as it would have the potential to mess up too many things.

Also, removed the `machine_type` field from the valid form data in `EditMachineFormTest`, for the same reason as mentioned above,
and added the missing `instance` kwarg.