Removed the unused `invert` tag
It might be wise to keep tags with such basic function as `invert`, even if they're not used, but since this tag outputs text for use with JavaScript,
it's likely better to write templates in a way that passes the un-inverted value to scripts, and let the inversion happen through JavaScript.
Also, the `invert` tag was usually easily replaced by the built-in `yesno` filter.