Enabled editing HTML source on internal home page
This is mainly for enabling adding tags like `<script>` and `<iframe>` that can be used to e.g. integrate MAKE's Google Calendar and Spotify playlist,
or show stats from the Minecraft server.

Also made it possible to edit the HTML source code of all content boxes (not just the internal home page) through Django admin,
as this is restricted to only a few members (currently leder, nestleder and the members of the Dev committee).

(It's probably only necessary to grant the members of the Dev committee the `internal.can_change_rich_text_source` permission.)