Added extra_change_permissions to ContentBox
Previously, it was (theoretically) possible to circumvent the permission restriction enforced by `EditInternalContentBoxView`, by simply visiting the
edit page of one of the content boxes on the main subdomain, and changing the PK in the URL to the one belonging to the internal home content box.
This commit fixes that loophole, and instead lets the site admins (staff) choose which permissions are required for editing a content box - no matter
what subdomain it's on - by setting the content boxes' `extra_change_permissions`.

Also removed the permission-related code from `EditInternalContentBoxView`, as the `internal.can_change_rich_text_source` permission
should be added to the internal home content box's `extra_change_permissions` instead.