Fixed only one language in content box forms
This could be reproduced by editing the internal content box, and then - without restarting the server - editing a content box on the main site
(e.g. the "about" content box), which would now also only have one language.

(For the record, I was not able to reproduce this bug on my machine, but three other members were able to.)