Removed EventTicket's legacy _name and _email
It's been a few years since we started requiring being logged in when registering for an event, and so it's definitely time to remove these backup
fields. (If it would ever be useful to dig up the ticket holders of old events, we have backups of the database from back then anyway.)

This change also allows for setting the `null` and `blank` options of the `user` field to `False` (by removing the kwargs),
and setting the `on_delete` option to `CASCADE`, which would be the standard way of handling these things.