Moved db queries from path converters to views
There was a change in Django 4.0 that made path converters run in an async context, which caused a `SynchronousOnlyOperation` error when querying
the database from a custom path converter (like subclasses of the previous `url_utils.SpecificObjectConverter`).
Using `sync_to_async()` did not work, but getting the view's "focused" object through path converters (instead of e.g. through views' `get_object()`
method) was an unconventional hack anyway, and so this was completely removed, and replaced by logic in the views.

It can also be noted that the previously mentioned error only happened when querying the database in the path converters' `to_python()` method,
not the `to_url()` method - where the former is usually invoked when Django receives a request and tries finding a view with a matching path,
and the latter is mainly invoked when reversing path names while constructing a response in views and templates.
This is why making database queries in `SpecificPageByTitle.to_url()` does not need to be changed.

Also removed the now unused `tests.utility.template_view_get_context_data()`.