Fixed 404 for docs main page in tests
This could be reproduced by running both `contentbox/tests/test_urls.py` (but not `test_views.py`) and the docs tests in the same test run.
Not entirely sure where and why the tests failed, but it's possible it had something to do with the previous code in `docs/urls.py` being run
when the module was imported in an earlier run test, and `Page.objects.get_or_create()` not being re-invoked when running the docs tests.