Removed reversing URLs using kwargs
Passing positional instead of keyword arguments, makes reversing the URL less error-prone, as one doesn't have to search through the whole codebase
to update the keyword argument names when the URL's path parameter is changed.

Exceptions to this include reversing e.g. `machine_detail`, as the URL contains multiple (and unconventionally named) parameters,
and it's easier to understand what the passed arguments are, if passed as keyword arguments.