Improved permission check for quotes
Now, users can delete and change quotes they've created, even if they don't have the `delete_quote` or `change_quote` permission.
To complement this, the quote permissions should be set up so that all members have the add permission,
but only privileged members have the delete and the "universal" change permission.

Also, the permission checks in `quotes_list.html` are now equal to the ones in the quote views.