Removed quote date blank and provided initial
Removed `blank`, as the field is not nullable, and we should therefore require the user to provide a value - e.g. the initial value in the view.
The latter is not strictly necessary, as the `SemanticDateInput` widget defaults to the current date if no initial value is provided, but setting
the initial value is best practice nonetheless.