Fixed MathJax filenames when using ManifestStaticFilesStorage
The filenames in the contents of the MathJax files were not changed to their hashed versions when running the `collectstatic` management command
(as should normally happen; see https://docs.djangoproject.com/en/4.0/ref/contrib/staticfiles/#manifeststaticfilesstorage),
which caused some JavaScript errors when the library tried to request its other files.

This could have been solved by replacing the names of files in the contents of the MathJax files, with their hashed versions,
but the way a lot of these filenames were being referenced to, made it hard to create patterns
(like the ones in `ManifestStaticFilesStorage.patterns`) for replacing these.

Placed all the MathJax files inside a folder with the version number as its name;
see the comments of the `MANIFEST_STATICFILES_IGNORE_PATTERNS` setting for reasoning.

Also moved the URL parameter for `MathJax.js` outside of the `static` template tag (in `mathjax_render.html`), as it's not part of the filename.