Fixed signal receivers being ignored
There was a bug that only seemed to happen on the server machine (running Ubuntu 20.04 and Python 3.9), that caused the signal receivers in
`groups/signals.py` and `internal/signals.py` (and presumably also `web/signals.py) to be disconnected at some point in time after connecting at
startup. I'm not sure why this did not happen on the CI servers (also running Ubuntu and testing Python 3.8-10) or my local machine, but it was found
out to be likely to have been caused by Python's interpreter garbage collecting the signal receivers, as they were defined as local functions inside
the scope of the `connect()` functions.