Improved equipment image layout
It now uses the raw image instead of a thumbnail, as it's not particularly necessary to use (storage space for) thumbnails when displaying just one
image, and it's often desirable as a user to be able to download the full-quality image - for whatever usage.

Also made the image width 600px (max) and made the height never exceed 50% of the viewport height.