Fixed redirect loop when logged in and unprivileged
I.e. when a logged-in user visited a page that required a permission that the user didn't have,
it would result in the user endlessly being redirected from the requested page, to the login page, and then back again.
This was caused by the `permission_required()` decorator (used in some apps' `urls.py`) redirecting users to the login page if they didn't have
the required permission (regardless of whether they were already logged in),
and the login view would redirect the users back to the originally requested page when users were already logged in.

This was solved by adding a new decorator, `permission_required_else_denied()`, that shows a `403 FORBIDDEN` page instead.

Also replaced most of the `permission_required()` decorator usages in `make_queue/urls.py` with making the views extend `PermissionRequiredMixin`
(this was already the case for most of the views, which meant that they had a double permission check),
as per the recommendation in the code review guideline (under "Improper access control (permissions)").