Fixed error message for multilingual widget
It was previously not showing, both due to the "language"/"languages" translations not being inserted into a JavaScript string,
which caused an exception, and because `event.preventDefault()` was not present,
which made the form submit the data the same instant as the error message was shown.

The call to `ngettext()` currently has no translation, as it's not recognized by `gettext` when outside of a `.js` file.
This will be fixed in an upcoming commit by extracting the JavaScript code to its own file.