Renamed advanced_course -> raise3d_course
This reflects the fact that the course in question is only really for Raise3D printers specifically.
It also reduces the potential for confusion when we later on add an SLA 3D printer course (which is technically also an "advanced" course).

Also set the `priority` of the "Workspaces" machine type to 100, like on the production website.