Split common_utils.js into two files
Keeping function definitions in `common_utils.js` while linking it without the `defer` keyword, makes it easier to use those functions in other
parts of the code that's not necessarily parsed after `common_utils.js` is when deferred.

The code in `common_deferred_code.js`, however, should always deferred, as the code it contains shouldn't be executed until after the DOM is ready.