Changed access list into course completion table
The information presented should be roughly the same, but the messages' focus is on course completion instead of access
(which is what the `if` statements check for), and it's now listed in a table with checkboxes instead of paragraphs.

Also added a contact message at the bottom of the profile page,
and moved the user's name to the top heading, to make it look better when the profile picture is not shown (for non-internal users).