Made pull request template always used
When it was placed within the `PULL_REQUEST_TEMPLATE` subdirectory, it was only possible to use the template through URL query parameters
(see the last bullet point under step 3,
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository#adding-a-pull-request-template),
which is not very usable to us.
Instead moving it to just below the `.github` directory, will make it so that the template is always used when opening a pull request.