Commented out unused code in ProfilePageView
Also removed the `image` template context variable, as it can simply be accessed through `profile` instead,

Also, removed the "Connect card number from course registration to user" code, which actually did nothing, as the lookup query checked the username
of an already connected user (`user__username`), and then simply overwrote the registration's user with the same user.
Furthermore, modifying the database is generally something that should be done in a GET request, and (working) code for connecting the card number
from a course registration to a user, already exists in more appropriate places in the codebase.