Made events' place more visible
The text is now bold, and the icon has been changed to a more traditional map marker icon (just with a hole in the middle).
Also added the same icon to the event lists.

Also did a small code refactor and removed the `line-height` property in `event_detail.css`
(as the `line-height: unset` property in `style.css` made it obsolete).