Optimized stream viewing on machine list page
Now, only the visible streams are connected to;
once the page is scrolled so that a stream image is no longer rendered, the stream for that machine is disconnected.

The machine detail page does not adopt this new behavior, as it's not really necessary when there is always only one stream on the page.