Improved error page for AdminEventTicketListView
...when trying to view the tickets connected to an `Event`/`TimePlace`, but the event is of the "wrong" type to view the tickets
 - i.e. viewing tickets connected to an `Event` that is of type `REPEATING`,
or viewing tickets connected to a `TimePlace` whose `event` is of type `STANDALONE`.

Viewing tickets when the event is of the "wrong" type is still possible if the `Event`/`TimePlace` has existing tickets connected to it,
as that would mean that the event type was changed after (some) people had registered for the `Event`/`TimePlace`,
and we shouldn't disallow viewing those existing tickets.

Also made the error code `404` instead of `403`.