Fixed unable to submit required search select widgets
...which are - among other places - used in the "Add New Member" form.

The standard widget attributes (through the `attrs.html` template) should generally always be placed on the "real" `<input>` tag holding the value
that will be submitted for the form field, and not the (usually dynamically created) `.search` input tag.

What seems to be a downside of this bugfix, is that the "Please fill out this field." error message (preventing the form from being submitted)
is no longer shown when trying to submit a form with a required field using this widget. Instead, an error message is shown after submitting the form
and the page is reloaded.
However, giving the `.search` input tag a `required` attribute (which is what enables the "Please fill out this field." error message) does not seem
to be the intended usage anyway, as it's not mentioned anywhere in the docs (https://fomantic-ui.com/modules/dropdown.html), and it causes this bug.