Made Profile.user non-nullable
The main reason for keeping the `SET_NULL` behavior, is that it would prevent voters of `SuggestSkill`s from being deleted,
but those suggested skills are not meant to be long-lived, and losing some votes out of the required 5 (the limit for "converting a `SuggestSkill` to
a `Skill`) has no impact on business.
Weighing heavier is not having to clean up profiles with deleted users, which is the main reason for this change.