Made member list change URL when viewing members
When the modal is shown after selecting a member in the list, the URL is now changed to the `member_detail` URL for that member,
and the member's name is prefixed to the page title (`<title>`). This adds/pushes an entry to the browser history.
All of this is reversed when the "Back" browser navigation button is pressed, or when the modal is hidden (by clicking outside of the modal
or hitting the "Esc" key) - which kind of simulates pressing the "Back" button.

This makes it easier to copy the URL for the current state of the page/modal,
avoids confusing the user when refreshing the page after e.g. having canceled editing a user and subsequently closing the modal,
and enables mobile (Android) users to close the modal by hitting the phone's "Back" button.

Known bug:
1. Open a new tab with the normal member list page (i.e. the URL `/members/`)
2. Select a member - so that the URL changes to e.g. `/members/1/`
3. Close the tab
4. Reopen the tab (using e.g. the Ctrl+Shift+T shortcut)
5. Make the modal hide (e.g. by hitting the "Esc" key)
The modal will close and immediately reopen.