Set minimum Python version to 3.10
It's time to bump the minimum required Python version now that 3.11 was recently released.
Also made it so that we only officially support the two latest versions (determined by the README and the versions tested in `python.yml`),
as it reduces the time necessary (by roughly a year by average - looking at the release dates of the last minor versions of Python 3)
before being able to use the newest language features - which is determined by the minimum required version, naturally.