Made Dataporten logout view use POST
...instead of GET.
This makes the logout button code in `header.html` easier, and removes the possibility of an adversary making users
modify their state by e.g. sending a link to https://makentnu.no/logout/, which would make the user log out when clicked;
this would not be possible now that the view uses POST with a CSRF token.

Also removed the `USES_DATAPORTEN_AUTH` context variable, as it's no longer used in any templates.