Fixed logout not working on subdomains
...other than the "main" subdomain (https://makentnu.no/).

When submitting the logout form on other subdomains (i.e. pressing the "Log out" button), multiple browsers (like Chrome) sent an `Origin` header with
a value of `null`, which caused the CSRF verification to fail. The browsers seem to do this because of their `Referrer Policy` being `same-origin`
when submitting the form (can be viewed e.g. through the "Network" tab in Chrome's developer tools) from one subdomain to another.

Also removed the `next` URL parameter from the logout form in `web/header.html`, both because Dataporten doesn't allow arbitrary logout redirect URLs,
and - mainly - because it's generally not that useful to be redirected back to the page after having logged out - especially if it's a page that
requires certain permissions, which would just cause the user to be immediately redirected to the login page after pressing the logout button.