Improved feedback of the toggle hiding occurrences
Now, the hide button (on each occurrence on an event admin page) changes from a closed, yellow eye to an open, grey eye
when changing the occurrence from hidden to shown. Also, the text in the popup that appears when hovering over
the button, now reflects the current `hidden` state.

Also refactored the code so that a minimal amount of styling is done in the JavaScript code (and in
`AdminNewsBaseToggleView`) - this is now done in the HTML code instead.

Also, fixed some usages of the deprecated `assertEquals()` method.