Moved most Django-related code to a new src folder
From the description of https://github.com/MAKENTNU/web/issues/601:
This makes it easier to get an overview over the role of the various files and folders in the project - especially if we
add e.g. top-level `docs`, `node` (already added), `docker` or similar folders in the future, which would currently be
difficult to spot among all the Django app folders.
It consequently also separates the Django code from other code.

Other changes necessary for / related to this move are added in the next commit.