Added button for delaying hiding secrets
This should make it less frustrating for users trying to see secrets containing larger amounts of text, and will
hopefully avoid creating situations where privileged users feel "forced" to visit the change form of a secret in order
to see it for longer, which defeats the whole purpose of making the secrets auto-hide themselves.

Also moved each secret's "hide" button (and "delay hiding" button) above the secret's content, to faster enable users to
hide the secret right after showing it, since the "show" and "hide" buttons now appear in the same spot.