Fixed faulty check logic in makemessages.py
The code didn't check if anything was actually changed before setting `changed` to `True`.
Also, adding an `if` statement checking whether each line was changed, is as inefficient - if not less - than doing one
check at the end (the latter of which was implemented in this commit) if the `.po` file already contained the correct
format of the file path comments. And if the latter is not the case, comparing the new and old file contents will return
relatively quickly once the comparison algorithm encounters the first file path comment of the file.