Removed explicit copyright end year from LICENSE
Constantly having to keep this year updated to the present year seems to simply be a long-standing norm
in the open source community, rather than something that's legally required.
There doesn't seem to be a consensus on exactly what to remove and what to keep, however,
so replacing the end year with "present" seems like the most conservative option that still removes the need
to constantly keep the year updated.

This was inspired by https://github.com/tox-dev/tox/pull/2941, which also links to a couple other "testimonies":
* https://daniel.haxx.se/blog/2023/01/08/copyright-without-years/
* https://hynek.me/til/copyright-years/