Reworked changing content boxes on other subdomains
Now, instead of returning a `404` HTTP status code, `ContentBoxUpdateView` redirects to the content box's change page
on the subdomain it "belongs on" (i.e. the subdomain that has a registered display URL for the content box).
Additionally, if the requested content box has no registered URL on any subdomain, the view will redirect to
the content box's Django admin change page.