Added urls, views and template for lore wiki
Added URLs and views for viewing the general lore article page and for viewing a particular lore article, for adding new lore article, and changing and deleting existing articles. Modified form_valid() method for the CreateView and the UpdateView, so the page throws an error if a user tries to write a new article on a topic that already exists.
Added new template for viewing lore articles, and added lore wiki to the internal header.