Removed constance database app from installed apps
...as mentioned in `django-constance`'s changelog for v3.0.0:
https://django-constance.readthedocs.io/en/latest/changes.html#v3-0-0-2023-07-27.

As a side note, I found it weird that the tests passed, even though this
non-existent app was enabled in the settings. However, through a little
bit of experimentation, I found out that it seems like Django doesn't
care whether the modules listed in `INSTALLED_APPS` are actually apps
- i.e. that they contain a file (named `apps.py` by default) containing
a subclass of `AppConfig`.