Changed validator for Discord username
Discord changed their usernames, and removed the hashtag and the four digits at the end of usernames. Discord usernames now consist of lowercase characters (a-z), numbers (0-9), underscores, and periods. Two consecutive periods are not allowed, and usernames have to be between 2 and 32 characters long. Updated the discord_username_regex accordingly.