Slight polish of the PR template
Made each checklist point written in the first person, which will
hopefully make PR authors more aware of the contents of each point,
and made them a little more succinct.

Also added PR references to the most recent changelog entries, to try to
establish a convention of doing so, as it will make it much easier to
find which PR made a certain change.