Moved and renamed Docker files
The name of the Docker Compose file now matches more closely with
the ones in the Docker docs, e.g. `compose.yaml` in https://docs.docker.com/compose/gettingstarted/#step-2-define-services-in-a-compose-file
and `compose.<something>.yaml` in https://docs.docker.com/compose/how-tos/multiple-compose-files/merge/
Chose a `.dev` suffix for it, to make it clearer that it's only used for
development stuff (not production).